+++ ges/gauss.h
@@ -25,14 +25,14 @@
 /* trivariate gaussian mixture */
 typedef struct gauss_mix_3d_t {
 	unsigned int mix_len; /* number of mixtures */
-	float *weight; /* weight coefficients */
+	double *weight; /* weight coefficients */
 	struct gauss_3d_t *each;
 } gauss_mix_3d_t;
 
 /* trivariate gaussian distribution */
 typedef struct gauss_3d_t {
-	float mean[3]; /* mean vector */
-	float covar[3][3]; /* covariance matrix */
+	double mean[3]; /* mean vector */
+	double covar[3][3]; /* covariance matrix */
 } gauss_3d_t;
 
 /* sample vector */
@@ -41,9 +41,9 @@ typedef struct sample_3d_t {
 } sample_3d_t;
 
 /* trivariate gaussian probability density function */
-float gauss_prob_den_3d(struct gauss_3d_t *gauss, struct sample_3d_t sample);
+double gauss_prob_den_3d(struct gauss_3d_t *gauss, struct sample_3d_t sample);
 /* trivariate gaussian discriminant function */
-float gauss_disc_3d(struct gauss_3d_t *gauss, struct sample_3d_t sample, float prior_prob);
+double gauss_disc_3d(struct gauss_3d_t *gauss, struct sample_3d_t sample, double prior_prob);
 /* populate with random numbers */
 void gauss_rand_3d(struct gauss_3d_t *gauss);
 /* print trivariate gaussian parameters to standard output */
@@ -54,9 +54,9 @@ void gauss_mix_create_3d(struct gauss_mix_3d_t *gauss_mix, unsigned int mix_len)
 /* de-allocate memory for each mixture */
 void gauss_mix_delete_3d(struct gauss_mix_3d_t *gauss_mix);
 /* trivariate gaussian mixture probability density function */
-float gauss_mix_prob_den_3d(struct gauss_mix_3d_t *gauss_mix, struct sample_3d_t sample);
+double gauss_mix_prob_den_3d(struct gauss_mix_3d_t *gauss_mix, struct sample_3d_t sample);
 /* trivariate gaussian mixture discriminant function */
-float gauss_mix_disc_3d(struct gauss_mix_3d_t *gauss_mix, struct sample_3d_t sample, float prior_prob);
+double gauss_mix_disc_3d(struct gauss_mix_3d_t *gauss_mix, struct sample_3d_t sample, double prior_prob);
 /* trivariate gaussian mixture density estimation */
 void gauss_mix_den_est_3d(struct gauss_mix_3d_t *gauss_mix, struct gauss_mix_3d_t *gauss_mix_est, struct sample_3d_t sample[], unsigned int sample_len);
 /* populate the trivariate gaussian mixture with random numbers */
